+++ src/actions/index.js
@@ -1,2 +1,3 @@
 export * from './authentication'
 export * from './configure'
+export * from './bookmarks'