+++ src/utils/tokenManager.js
@@ -1,3 +1,4 @@
+import { ID, TOKEN, AUTH_INFO } from '../constants/localStorage'
 import get from 'lodash/get'
 
 const _ = {
@@ -5,6 +6,8 @@ const _ = {
 }
 
 let tokenRenewalTimeout
+
+// Internal Usage
 
 const daysToSecs = (days) => {
   if (typeof days === 'number') {
@@ -34,6 +37,8 @@ function localStorageExist(callback) {
   }
   return null
 }
+
+// External Usage
 
 /**
 * @param {object} setting - inside obj should be key value pair for localStorage
@@ -99,4 +104,32 @@ export function scheduleRenewToken(days, cb) {
   tokenRenewalTimeout = setInterval(() => {
     cb()
   }, daysToSecs(days))
+}
+
+
+const getVerifiedToken = () => {
+  const browserLocalStorage = (typeof localStorage === 'undefined') ? null : localStorage
+  if (browserLocalStorage) {
+    const authInfoString = browserLocalStorage.getItem(AUTH_INFO)
+    const authInfoObj = JSON.parse(authInfoString)
+    return _.get(authInfoObj, TOKEN, '')
+  }
+  return null
+}
+
+const getUserId = () => {
+  const browserLocalStorage = (typeof localStorage === 'undefined') ? null : localStorage
+  if (browserLocalStorage) {
+    const authInfoString = browserLocalStorage.getItem(AUTH_INFO)
+    const authInfoObj = JSON.parse(authInfoString)
+    return _.get(authInfoObj, ID, '')
+  }
+  return null
+}
+
+export const getNecessaryInfo = () => {
+  return {
+    token: getVerifiedToken(),
+    userId: getUserId(),
+  }
 }